Refactor for code style, generic component use, and better UI presentation
