+++ src/features/basic/index.ts
@@ -77,4 +77,4 @@ import './sidebar-contracts-details/sidebar-contracts-details';
 import './table-rows-alternating-colors';
 import './tile-controls-background';
 import './usr-subscription-level';
-import './prunstats-shortcut/prunstats-shortcut.ts';
+import './prunstats-shortcut';
